+++ ShareX.ScreenCaptureLib/Forms/ScrollingCaptureLightForm.cs
@@ -25,11 +25,7 @@
 
 using ShareX.HelpersLib;
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Imaging;
-using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
@@ -41,13 +37,8 @@ namespace ShareX.ScreenCaptureLib
         public event Action<Bitmap> UploadRequested;
 
         public ScrollingCaptureOptions Options { get; private set; }
-        public Bitmap Result { get; private set; }
 
-        private List<Bitmap> images = new List<Bitmap>();
-        private bool isCapturing, stopRequested;
-        private int currentScrollCount, bestMatchCount, bestMatchIndex;
-        private WindowInfo selectedWindow;
-        private Rectangle selectedRectangle;
+        private ScrollingCaptureManager manager;
         private Point dragStartPosition;
 
         public ScrollingCaptureLightForm(ScrollingCaptureOptions options)
@@ -56,6 +47,8 @@ namespace ShareX.ScreenCaptureLib
 
             InitializeComponent();
             ShareXResources.ApplyTheme(this);
+
+            manager = new ScrollingCaptureManager(Options);
         }
 
         protected override void Dispose(bool disposing)
@@ -63,35 +56,10 @@ namespace ShareX.ScreenCaptureLib
             if (disposing)
             {
                 components?.Dispose();
-
-                Reset();
+                manager?.Dispose();
             }
 
             base.Dispose(disposing);
-        }
-
-        private void Reset()
-        {
-            currentScrollCount = 0;
-
-            ResetPictureBox();
-            btnUpload.Enabled = false;
-
-            if (images != null)
-            {
-                foreach (Bitmap bmp in images)
-                {
-                    bmp?.Dispose();
-                }
-
-                images.Clear();
-            }
-
-            if (Result != null)
-            {
-                Result.Dispose();
-                Result = null;
-            }
         }
 
         private void ResetPictureBox()
@@ -103,121 +71,21 @@ namespace ShareX.ScreenCaptureLib
 
         private async Task StartCapture()
         {
-            if (!isCapturing)
+            if (!manager.IsCapturing)
             {
-                isCapturing = true;
-                stopRequested = false;
-                bestMatchCount = 0;
-                bestMatchIndex = 0;
-
-                btnCapture.Enabled = false;
                 WindowState = FormWindowState.Minimized;
-                Reset();
-                selectedWindow.Activate();
+                btnCapture.Enabled = false;
+                btnUpload.Enabled = false;
+                ResetPictureBox();
 
-                await Task.Delay(Options.StartDelay);
-                await CaptureJob();
+                await manager.StartCapture();
+
+                btnCapture.Enabled = true;
+                btnUpload.Enabled = manager.Result != null;
+                pbOutput.Image = manager.Result;
+
+                this.ForceActivate();
             }
-        }
-
-        private void StopCapture()
-        {
-            if (isCapturing)
-            {
-                stopRequested = true;
-            }
-        }
-
-        private void EndCapture()
-        {
-            btnCapture.Enabled = true;
-            btnUpload.Enabled = Result != null;
-            pbOutput.Image = Result;
-
-            isCapturing = false;
-
-            this.ForceActivate();
-        }
-
-        private async Task CaptureJob()
-        {
-            InputHelpers.SendKeyPress(VirtualKeyCode.HOME);
-            NativeMethods.SendMessage(selectedWindow.Handle, (int)WindowsMessages.VSCROLL, (int)ScrollBarCommands.SB_TOP, 0);
-
-            Stopwatch timer = new Stopwatch();
-
-            do
-            {
-                int delay = Options.ScrollDelay - (int)timer.ElapsedMilliseconds;
-
-                if (delay > 0)
-                {
-                    await Task.Delay(delay);
-                }
-
-                if (stopRequested)
-                {
-                    break;
-                }
-
-                Screenshot screenshot = new Screenshot()
-                {
-                    CaptureCursor = false
-                };
-
-                Bitmap bmp = screenshot.CaptureRectangle(selectedRectangle);
-
-                if (bmp != null)
-                {
-                    images.Add(bmp);
-                }
-
-                InputHelpers.SendMouseWheel(-120 * 2);
-                currentScrollCount++;
-
-                timer.Restart();
-
-                if (images.Count > 0)
-                {
-                    Result = await CombineImagesAsync(Result, images[images.Count - 1]);
-                }
-            }
-            while (!stopRequested && currentScrollCount <= Options.MaximumScrollCount && !IsScrollReachedBottom(selectedWindow.Handle));
-
-            EndCapture();
-        }
-
-        private bool IsScrollReachedBottom(IntPtr handle)
-        {
-            SCROLLINFO scrollInfo = new SCROLLINFO();
-            scrollInfo.cbSize = (uint)Marshal.SizeOf(scrollInfo);
-            scrollInfo.fMask = (uint)(ScrollInfoMask.SIF_RANGE | ScrollInfoMask.SIF_PAGE | ScrollInfoMask.SIF_TRACKPOS);
-
-            if (NativeMethods.GetScrollInfo(handle, (int)SBOrientation.SB_VERT, ref scrollInfo))
-            {
-                return scrollInfo.nMax == scrollInfo.nTrackPos + scrollInfo.nPage - 1;
-            }
-
-            return IsLastTwoImagesSame();
-        }
-
-        private bool IsLastTwoImagesSame()
-        {
-            bool result = false;
-
-            if (images.Count > 1)
-            {
-                result = ImageHelpers.CompareImages(images[images.Count - 1], images[images.Count - 2]);
-
-                if (result)
-                {
-                    Bitmap last = images[images.Count - 1];
-                    images.Remove(last);
-                    last.Dispose();
-                }
-            }
-
-            return result;
         }
 
         private async Task SelectWindow()
@@ -225,7 +93,7 @@ namespace ShareX.ScreenCaptureLib
             WindowState = FormWindowState.Minimized;
             Thread.Sleep(250);
 
-            if (RegionCaptureTasks.GetRectangleRegion(out selectedRectangle, out selectedWindow, new RegionCaptureOptions()))
+            if (manager.SelectWindow())
             {
                 await StartCapture();
             }
@@ -235,106 +103,11 @@ namespace ShareX.ScreenCaptureLib
             }
         }
 
-        private async Task<Bitmap> CombineImagesAsync(Bitmap result, Bitmap currentImage)
-        {
-            return await Task.Run(() => CombineImages(result, currentImage));
-        }
-
-        private Bitmap CombineImages(Bitmap result, Bitmap currentImage)
-        {
-            if (result == null)
-            {
-                return (Bitmap)currentImage.Clone();
-            }
-
-            int matchCount = 0;
-            int matchIndex = 0;
-            int matchLimit = currentImage.Height / 2;
-            int ignoreSideOffset = Math.Max(50, currentImage.Width / 20);
-
-            if (currentImage.Width < ignoreSideOffset * 3)
-            {
-                ignoreSideOffset = 0;
-            }
-
-            Rectangle rect = new Rectangle(ignoreSideOffset, result.Height - currentImage.Height, currentImage.Width - ignoreSideOffset * 2, currentImage.Height);
-
-            BitmapData bdResult = result.LockBits(new Rectangle(0, 0, result.Width, result.Height), ImageLockMode.ReadOnly, PixelFormat.Format32bppArgb);
-            BitmapData bdCurrentImage = currentImage.LockBits(new Rectangle(0, 0, currentImage.Width, currentImage.Height), ImageLockMode.ReadOnly, PixelFormat.Format32bppArgb);
-            int stride = bdResult.Stride;
-            int pixelSize = stride / result.Width;
-            IntPtr resultScan0 = bdResult.Scan0 + pixelSize * ignoreSideOffset;
-            IntPtr currentImageScan0 = bdCurrentImage.Scan0 + pixelSize * ignoreSideOffset;
-            int rectBottom = rect.Bottom - 1;
-            int compareLength = pixelSize * rect.Width;
-
-            for (int currentImageY = currentImage.Height - 1; currentImageY >= 0 && matchCount < matchLimit; currentImageY--)
-            {
-                int currentMatchCount = 0;
-
-                for (int y = 0; currentImageY - y >= 0 && currentMatchCount < matchLimit; y++)
-                {
-                    if (NativeMethods.memcmp(resultScan0 + ((rectBottom - y) * stride), currentImageScan0 + ((currentImageY - y) * stride), compareLength) == 0)
-                    {
-                        currentMatchCount++;
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-
-                if (currentMatchCount > matchCount)
-                {
-                    matchCount = currentMatchCount;
-                    matchIndex = currentImageY;
-                }
-            }
-
-            result.UnlockBits(bdResult);
-            currentImage.UnlockBits(bdCurrentImage);
-
-            if (matchCount == 0 && bestMatchCount > 0)
-            {
-                matchCount = bestMatchCount;
-                matchIndex = bestMatchIndex;
-            }
-
-            if (matchCount > 0)
-            {
-                int matchHeight = currentImage.Height - matchIndex - 1;
-
-                if (matchHeight > 0)
-                {
-                    if (matchCount > bestMatchCount)
-                    {
-                        bestMatchCount = matchCount;
-                        bestMatchIndex = matchIndex;
-                    }
-
-                    Bitmap newResult = new Bitmap(result.Width, result.Height + matchHeight);
-
-                    using (Graphics g = Graphics.FromImage(newResult))
-                    {
-                        g.DrawImage(result, new Rectangle(0, 0, result.Width, result.Height),
-                            new Rectangle(0, 0, result.Width, result.Height), GraphicsUnit.Pixel);
-                        g.DrawImage(currentImage, new Rectangle(0, result.Height, currentImage.Width, matchHeight),
-                            new Rectangle(0, matchIndex + 1, currentImage.Width, matchHeight), GraphicsUnit.Pixel);
-                    }
-
-                    result.Dispose();
-                    result = newResult;
-                }
-            }
-
-            return result;
-        }
-
         private void UploadResult()
         {
-            if (Result != null)
+            if (manager.Result != null)
             {
-                OnUploadRequested((Bitmap)Result.Clone());
+                OnUploadRequested((Bitmap)manager.Result.Clone());
             }
         }
 
@@ -350,7 +123,7 @@ namespace ShareX.ScreenCaptureLib
 
         private void ScrollingCaptureLightForm_Activated(object sender, EventArgs e)
         {
-            StopCapture();
+            manager.StopCapture();
         }
 
         private async void btnCapture_Click(object sender, EventArgs e)
