+++ ShareX.ScreenCaptureLib/Helpers/ImageEditorButton.cs
