+++ ShareX.ScreenCaptureLib/Helpers/ImageEditorControl.cs
