+++ ShareX.ScreenCaptureLib/Helpers/ImageEditorScrollbar.cs
