+++ ShareX.ScreenCaptureLib/Helpers/InputManager.cs
