+++ ShareX.ScreenCaptureLib/Helpers/LocationInfo.cs
