+++ ShareX.ScreenCaptureLib/Helpers/MouseState.cs
