+++ ShareX.ScreenCaptureLib/Helpers/ResizeNode.cs
