+++ ShareX.ScreenCaptureLib/Helpers/ScrollbarManager.cs
