+++ ShareX.ScreenCaptureLib/Helpers/ScrollingCaptureManager.cs
@@ -1,0 +1,283 @@
+﻿#region License Information (GPL v3)
+
+/*
+    ShareX - A program that allows you to take screenshots and share any file type
+    Copyright (c) 2007-2023 ShareX Team
+
+    This program is free software; you can redistribute it and/or
+    modify it under the terms of the GNU General Public License
+    as published by the Free Software Foundation; either version 2
+    of the License, or (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program; if not, write to the Free Software
+    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+    Optionally you can also view the license at <http://www.gnu.org/licenses/>.
+*/
+
+#endregion License Information (GPL v3)
+
+using ShareX.HelpersLib;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Drawing;
+using System.Drawing.Imaging;
+using System.Runtime.InteropServices;
+using System.Threading.Tasks;
+
+namespace ShareX.ScreenCaptureLib
+{
+    internal class ScrollingCaptureManager : IDisposable
+    {
+        public ScrollingCaptureOptions Options { get; private set; }
+        public Bitmap Result { get; private set; }
+        public bool IsCapturing { get; private set; }
+
+        private List<Bitmap> images = new List<Bitmap>();
+        private bool stopRequested;
+        private int bestMatchCount, bestMatchIndex, currentScrollCount;
+        private WindowInfo selectedWindow;
+        private Rectangle selectedRectangle;
+
+        public ScrollingCaptureManager(ScrollingCaptureOptions options)
+        {
+            Options = options;
+        }
+
+        public void Dispose()
+        {
+            if (images != null)
+            {
+                foreach (Bitmap bmp in images)
+                {
+                    bmp?.Dispose();
+                }
+
+                images.Clear();
+            }
+
+            if (Result != null)
+            {
+                Result.Dispose();
+                Result = null;
+            }
+        }
+
+        private void Reset()
+        {
+            IsCapturing = false;
+            stopRequested = false;
+            bestMatchCount = 0;
+            bestMatchIndex = 0;
+            currentScrollCount = 0;
+
+            Dispose();
+        }
+
+        public async Task StartCapture()
+        {
+            if (!IsCapturing && selectedWindow != null)
+            {
+                Reset();
+
+                IsCapturing = true;
+
+                selectedWindow.Activate();
+
+                await Task.Delay(Options.StartDelay);
+
+                InputHelpers.SendKeyPress(VirtualKeyCode.HOME);
+                NativeMethods.SendMessage(selectedWindow.Handle, (int)WindowsMessages.VSCROLL, (int)ScrollBarCommands.SB_TOP, 0);
+
+                Stopwatch timer = new Stopwatch();
+
+                do
+                {
+                    int delay = Options.ScrollDelay - (int)timer.ElapsedMilliseconds;
+
+                    if (delay > 0)
+                    {
+                        await Task.Delay(delay);
+                    }
+
+                    if (stopRequested)
+                    {
+                        break;
+                    }
+
+                    Screenshot screenshot = new Screenshot()
+                    {
+                        CaptureCursor = false
+                    };
+
+                    Bitmap bmp = screenshot.CaptureRectangle(selectedRectangle);
+
+                    if (bmp != null)
+                    {
+                        images.Add(bmp);
+                    }
+
+                    InputHelpers.SendMouseWheel(-120 * 2);
+                    currentScrollCount++;
+                    timer.Restart();
+
+                    if (images.Count > 0)
+                    {
+                        Result = await CombineImagesAsync(Result, images[images.Count - 1]);
+                    }
+                }
+                while (!stopRequested && currentScrollCount <= Options.MaximumScrollCount && !IsScrollReachedBottom(selectedWindow.Handle));
+
+                IsCapturing = false;
+            }
+        }
+
+        public void StopCapture()
+        {
+            if (IsCapturing)
+            {
+                stopRequested = true;
+            }
+        }
+
+        public bool SelectWindow()
+        {
+            return RegionCaptureTasks.GetRectangleRegion(out selectedRectangle, out selectedWindow, new RegionCaptureOptions());
+        }
+
+        private bool IsScrollReachedBottom(IntPtr handle)
+        {
+            SCROLLINFO scrollInfo = new SCROLLINFO();
+            scrollInfo.cbSize = (uint)Marshal.SizeOf(scrollInfo);
+            scrollInfo.fMask = (uint)(ScrollInfoMask.SIF_RANGE | ScrollInfoMask.SIF_PAGE | ScrollInfoMask.SIF_TRACKPOS);
+
+            if (NativeMethods.GetScrollInfo(handle, (int)SBOrientation.SB_VERT, ref scrollInfo))
+            {
+                return scrollInfo.nMax == scrollInfo.nTrackPos + scrollInfo.nPage - 1;
+            }
+
+            return IsLastTwoImagesSame();
+        }
+
+        private bool IsLastTwoImagesSame()
+        {
+            bool result = false;
+
+            if (images.Count > 1)
+            {
+                result = ImageHelpers.CompareImages(images[images.Count - 1], images[images.Count - 2]);
+
+                if (result)
+                {
+                    Bitmap last = images[images.Count - 1];
+                    images.Remove(last);
+                    last.Dispose();
+                }
+            }
+
+            return result;
+        }
+
+        private async Task<Bitmap> CombineImagesAsync(Bitmap result, Bitmap currentImage)
+        {
+            return await Task.Run(() => CombineImages(result, currentImage));
+        }
+
+        private Bitmap CombineImages(Bitmap result, Bitmap currentImage)
+        {
+            if (result == null)
+            {
+                return (Bitmap)currentImage.Clone();
+            }
+
+            int matchCount = 0;
+            int matchIndex = 0;
+            int matchLimit = currentImage.Height / 2;
+            int ignoreSideOffset = Math.Max(50, currentImage.Width / 20);
+
+            if (currentImage.Width < ignoreSideOffset * 3)
+            {
+                ignoreSideOffset = 0;
+            }
+
+            Rectangle rect = new Rectangle(ignoreSideOffset, result.Height - currentImage.Height, currentImage.Width - ignoreSideOffset * 2, currentImage.Height);
+
+            BitmapData bdResult = result.LockBits(new Rectangle(0, 0, result.Width, result.Height), ImageLockMode.ReadOnly, PixelFormat.Format32bppArgb);
+            BitmapData bdCurrentImage = currentImage.LockBits(new Rectangle(0, 0, currentImage.Width, currentImage.Height), ImageLockMode.ReadOnly, PixelFormat.Format32bppArgb);
+            int stride = bdResult.Stride;
+            int pixelSize = stride / result.Width;
+            IntPtr resultScan0 = bdResult.Scan0 + pixelSize * ignoreSideOffset;
+            IntPtr currentImageScan0 = bdCurrentImage.Scan0 + pixelSize * ignoreSideOffset;
+            int rectBottom = rect.Bottom - 1;
+            int compareLength = pixelSize * rect.Width;
+
+            for (int currentImageY = currentImage.Height - 1; currentImageY >= 0 && matchCount < matchLimit; currentImageY--)
+            {
+                int currentMatchCount = 0;
+
+                for (int y = 0; currentImageY - y >= 0 && currentMatchCount < matchLimit; y++)
+                {
+                    if (NativeMethods.memcmp(resultScan0 + ((rectBottom - y) * stride), currentImageScan0 + ((currentImageY - y) * stride), compareLength) == 0)
+                    {
+                        currentMatchCount++;
+                    }
+                    else
+                    {
+                        break;
+                    }
+                }
+
+                if (currentMatchCount > matchCount)
+                {
+                    matchCount = currentMatchCount;
+                    matchIndex = currentImageY;
+                }
+            }
+
+            result.UnlockBits(bdResult);
+            currentImage.UnlockBits(bdCurrentImage);
+
+            if (matchCount == 0 && bestMatchCount > 0)
+            {
+                matchCount = bestMatchCount;
+                matchIndex = bestMatchIndex;
+            }
+
+            if (matchCount > 0)
+            {
+                int matchHeight = currentImage.Height - matchIndex - 1;
+
+                if (matchHeight > 0)
+                {
+                    if (matchCount > bestMatchCount)
+                    {
+                        bestMatchCount = matchCount;
+                        bestMatchIndex = matchIndex;
+                    }
+
+                    Bitmap newResult = new Bitmap(result.Width, result.Height + matchHeight);
+
+                    using (Graphics g = Graphics.FromImage(newResult))
+                    {
+                        g.DrawImage(result, new Rectangle(0, 0, result.Width, result.Height),
+                            new Rectangle(0, 0, result.Width, result.Height), GraphicsUnit.Pixel);
+                        g.DrawImage(currentImage, new Rectangle(0, result.Height, currentImage.Width, matchHeight),
+                            new Rectangle(0, matchIndex + 1, currentImage.Width, matchHeight), GraphicsUnit.Pixel);
+                    }
+
+                    result.Dispose();
+                    result = newResult;
+                }
+            }
+
+            return result;
+        }
+    }
+}
