+++ ShareX.ScreenCaptureLib/Helpers/SimpleWindowInfo.cs
