+++ ShareX.ScreenCaptureLib/Helpers/SnapSize.cs
