+++ ShareX.ScreenCaptureLib/Helpers/WindowsList.cs
