+++ ShareX.ScreenCaptureLib/Helpers/WindowsRectangleList.cs
