+++ ShareX.ScreenCaptureLib/ShareX.ScreenCaptureLib.csproj
@@ -140,9 +140,10 @@
     <Compile Include="Forms\StickerPackForm.Designer.cs">
       <DependentUpon>StickerPackForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="RegionHelpers\ImageEditorButton.cs" />
-    <Compile Include="RegionHelpers\ImageEditorScrollbar.cs" />
-    <Compile Include="RegionHelpers\ScrollbarManager.cs" />
+    <Compile Include="Helpers\ImageEditorButton.cs" />
+    <Compile Include="Helpers\ImageEditorScrollbar.cs" />
+    <Compile Include="Helpers\ScrollbarManager.cs" />
+    <Compile Include="Helpers\ScrollingCaptureManager.cs" />
     <Compile Include="Shapes\AnnotationOptions.cs" />
     <Compile Include="Enums.cs" />
     <Compile Include="Forms\RegionCaptureTransparentForm.cs">
@@ -206,24 +207,24 @@
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="ScreenRecording\ImageCache.cs" />
-    <Compile Include="RegionHelpers\LocationInfo.cs" />
+    <Compile Include="Helpers\LocationInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="Shapes\ShapeManager.cs" />
-    <Compile Include="RegionHelpers\ImageEditorControl.cs" />
+    <Compile Include="Helpers\ImageEditorControl.cs" />
     <Compile Include="Forms\RegionCaptureForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="RegionHelpers\InputManager.cs" />
-    <Compile Include="RegionHelpers\MouseState.cs" />
-    <Compile Include="RegionHelpers\ResizeNode.cs" />
+    <Compile Include="Helpers\InputManager.cs" />
+    <Compile Include="Helpers\MouseState.cs" />
+    <Compile Include="Helpers\ResizeNode.cs" />
     <Compile Include="RegionCaptureTasks.cs" />
     <Compile Include="ScrollingCaptureOptions.cs" />
-    <Compile Include="RegionHelpers\SimpleWindowInfo.cs" />
-    <Compile Include="RegionHelpers\SnapSize.cs" />
+    <Compile Include="Helpers\SimpleWindowInfo.cs" />
+    <Compile Include="Helpers\SnapSize.cs" />
     <Compile Include="RegionCaptureOptions.cs" />
     <Compile Include="ScreenRecording\ScreenRecorder.cs" />
     <Compile Include="ScreenRecording\HardDiskCache.cs" />
@@ -231,8 +232,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Screenshot_Transparent.cs" />
     <Compile Include="Animations\TextAnimation.cs" />
-    <Compile Include="RegionHelpers\WindowsList.cs" />
-    <Compile Include="RegionHelpers\WindowsRectangleList.cs" />
+    <Compile Include="Helpers\WindowsList.cs" />
+    <Compile Include="Helpers\WindowsRectangleList.cs" />
     <Compile Include="Shapes\Tool\CutOutTool.cs" />
   </ItemGroup>
   <ItemGroup>
